Moved cWorld manual bindings out into a separate file.
