+++ src/Bindings/LuaState.cpp
@@ -107,7 +107,7 @@ void cLuaState::Create(void)
 void cLuaState::RegisterAPILibs(void)
 {
 	tolua_AllToLua_open(m_LuaState);
-	ManualBindings::Bind(m_LuaState);
+	cManualBindings::Bind(m_LuaState);
 	DeprecatedBindings::Bind(m_LuaState);
 	luaopen_lsqlite3(m_LuaState);
 	luaopen_lxp(m_LuaState);
